add files that show that there are 2 bugs in the Dijkstra function
From reading the "two_bugs.txt" file, you can say that the minimum
path is 18, but the program says a different thing.

The fix is left as an exercise to the reader.

I added the file "single_bug.txt" because I think there are 2 bugs :)

"single_bug.txt"  only triggers one bug, so you will know
when you have fixed it, and then you can move on to debugging "two_bugs.txt".